utils/firestore.js utils/firestoreConfig.js components/TextInput.js contexts/SigninContext.js
• Firestore connection was abstracted to its own file to keep the
DRY approach
----------
components/app/GearIcon.js
components/app/Modal.js
components/app/TokenInstructions.js
components/app/TokenSaved.js
pages/app.js

• Notion token configuration page completed
