+++ components/TextInput.js
@@ -3,20 +3,9 @@ import {
     FirestoreProvider,
     FirestoreMutation,
 } from "@react-firebase/firestore";
-import "firebase/firestore";
-import firebase from "firebase/app";
 import { EmailContext } from "../contexts/EmailContext";
 import SubmitButton from "./SubmitButton";
-
-const config = {
-    apiKey: process.env.NEXT_PUBLIC_API_KEY,
-    authDomain: process.env.NEXT_PUBLIC_AUTH_DOMAIN,
-    projectId: process.env.NEXT_PUBLIC_PROJECT_ID,
-    storageBucket: process.env.NEXT_PUBLIC_STORAGE_BUCKET,
-    messagingSenderId: process.env.NEXT_PUBLIC_MESSAGING_SENDER_ID,
-    appId: process.env.NEXT_PUBLIC_APP_ID,
-    measurementId: process.env.NEXT_PUBLIC_MEASUREMENT_ID,
-};
+import config from "../utils/firestoreConfig";
 
 export default function TextInput() {
     const { ValidateAndSubmit } = useContext(EmailContext);
